Updated repository for full expanded dataset

+++ create_data.sh
@@ -2,29 +2,33 @@
 
 # This is the script to generate the datasets
 idir=`pwd`
-mkdir wdir && cd wdir
+mkdir -p wdir && cd wdir
 
 # Loop over different numbers of nucleosomes
-for nnucl in {5..100..5}; do
-    mkdir $nnucl && cd $nnucl
+for nnucl in {2..100..1}; do
+    mkdir -p $nnucl
+    cd $nnucl
 
     # Loop over 10 replicas
     for irep in {0..9..1}; do
-        mkdir $irep && cd $irep
+        mkdir -p $irep
+        cd $irep
 
-        # Copy seed files here
-        cp $idir/seed/* .
+        # If file doesn't exist, make new data
+        if [ ! -f traj.dump ]; then
+            # Copy seed files here
+            cp $idir/seed/* .
 
-        # Change the random number initializer
-        cmd="sed -i -e 's/RANDOM/$RANDOM/g' in.liquidcrystal"
-        eval $cmd
+            # Change the random number initializer
+            cmd="sed -i -e 's/RANDOM/$RANDOM/g' in.nucleosomes"
+            eval $cmd
 
-        # Run the initialization script
-        python2 $idir/init/init_1cpn_nucleosomes.py -n $nnucl
+            # Run the initialization script
+            python2 $idir/init/init_1cpn_nucleosomes.py -n $nnucl
 
-        # Run lammps
-        lmp_mpi < in.nucleosomes
-
+            # Run lammps
+            lmp_mpi < in.nucleosomes
+        fi
         cd ..
     done
